create two new ways to generate alu and jmp operations
