+++ pkg/ebpf/BUILD
@@ -27,7 +27,8 @@ go_library(
         "base_structures.go",
         "constants.go",
         "encoding_functions.go",
-        "generation_utils.go",
+        "instruction_generators.go",
+        "instruction_sequence.go",
         "jmp_instructions.go",
         "poc_generator.go",
         "program.go",
